A Dockerfile, requirements.txt file and a Jupyter notebook ready to use!
This is the data behind the data analysis for one of my earliest medium blog posts. These files are ready to go: if you have Docker, feel free to download the files and run the dockerfile. If you want to run the notebook on your local machine then just download the jupyter notebook file and/or the .txt file based on what modules you have in your local environment.

https://medium.com/@m.adeleke1/are-running-backs-overvalued-in-todays-nfl-a-chatgpt-analysis-eecee58e1353.